Move SupportUs page to RadicalCollaboration (#596) (#614)
* Move SupportUs page files to RadicalCollaboration directory

* Move RadicalCollaboration subpages to subdirectories

* Clean up App.js

* Update some routes/route references
