+++ src/pages/RadicalCollaboration/FAQCardDesktop.js
