+++ src/pages/RadicalCollaboration/PhotoCardDesktop.js
